:truck: chore(api) move `verify-certification-code-service` to `src`
